third_party/blink/web_tests/platform/generic/virtual/origin-trials-runtimeflags-disabled/http/tests/origin_trials/webexposed/animationworklet-origin-trial-interfaces-worklet-scope-expected.txt

CONSOLE MESSAGE: line 13: List all interfaces in the animation worklet global scope.
CONSOLE MESSAGE: line 153: interface AnimationWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE: line 153:     getter highWaterMark
CONSOLE MESSAGE: line 153:     getter size
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface CountQueuingStrategy
CONSOLE MESSAGE: line 153:     getter highWaterMark
CONSOLE MESSAGE: line 153:     getter size
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface ReadableStream
CONSOLE MESSAGE: line 153:     getter locked
CONSOLE MESSAGE: line 153:     method cancel
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method getReader
CONSOLE MESSAGE: line 153:     method pipeThrough
CONSOLE MESSAGE: line 153:     method pipeTo
CONSOLE MESSAGE: line 153:     method tee
CONSOLE MESSAGE: line 153: interface ReadableStreamDefaultReader
CONSOLE MESSAGE: line 153:     getter closed
CONSOLE MESSAGE: line 153:     method cancel
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method read
CONSOLE MESSAGE: line 153:     method releaseLock
CONSOLE MESSAGE: line 153: interface TransformStream
CONSOLE MESSAGE: line 153:     getter readable
CONSOLE MESSAGE: line 153:     getter writable
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface WorkletAnimationEffect
CONSOLE MESSAGE: line 153:     getter localTime
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method getComputedTiming
CONSOLE MESSAGE: line 153:     method getTiming
CONSOLE MESSAGE: line 153:     setter localTime
CONSOLE MESSAGE: line 153: interface WorkletGlobalScope
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface WritableStream
CONSOLE MESSAGE: line 153:     getter locked
CONSOLE MESSAGE: line 153:     method abort
CONSOLE MESSAGE: line 153:     method close
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method getWriter
CONSOLE MESSAGE: line 153: interface WritableStreamDefaultWriter
CONSOLE MESSAGE: line 153:     getter closed
CONSOLE MESSAGE: line 153:     getter desiredSize
CONSOLE MESSAGE: line 153:     getter ready
CONSOLE MESSAGE: line 153:     method abort
CONSOLE MESSAGE: line 153:     method close
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method releaseLock
CONSOLE MESSAGE: line 153:     method write
CONSOLE MESSAGE: line 153: global object
CONSOLE MESSAGE: line 153:     attribute console
CONSOLE MESSAGE: line 153:     attribute globalThis
CONSOLE MESSAGE: line 153:     method gc
CONSOLE MESSAGE: line 153:     method registerAnimator
CONSOLE MESSAGE: line 153: interface AnimationWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE: line 153:     getter highWaterMark
CONSOLE MESSAGE: line 153:     getter size
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface CountQueuingStrategy
CONSOLE MESSAGE: line 153:     getter highWaterMark
CONSOLE MESSAGE: line 153:     getter size
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface ReadableStream
CONSOLE MESSAGE: line 153:     getter locked
CONSOLE MESSAGE: line 153:     method cancel
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method getReader
CONSOLE MESSAGE: line 153:     method pipeThrough
CONSOLE MESSAGE: line 153:     method pipeTo
CONSOLE MESSAGE: line 153:     method tee
CONSOLE MESSAGE: line 153: interface ReadableStreamDefaultReader
CONSOLE MESSAGE: line 153:     getter closed
CONSOLE MESSAGE: line 153:     method cancel
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method read
CONSOLE MESSAGE: line 153:     method releaseLock
CONSOLE MESSAGE: line 153: interface TransformStream
CONSOLE MESSAGE: line 153:     getter readable
CONSOLE MESSAGE: line 153:     getter writable
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface WorkletAnimationEffect
CONSOLE MESSAGE: line 153:     getter localTime
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method getComputedTiming
CONSOLE MESSAGE: line 153:     method getTiming
CONSOLE MESSAGE: line 153:     setter localTime
CONSOLE MESSAGE: line 153: interface WorkletGlobalScope
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153: interface WritableStream
CONSOLE MESSAGE: line 153:     getter locked
CONSOLE MESSAGE: line 153:     method abort
CONSOLE MESSAGE: line 153:     method close
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method getWriter
CONSOLE MESSAGE: line 153: interface WritableStreamDefaultWriter
CONSOLE MESSAGE: line 153:     getter closed
CONSOLE MESSAGE: line 153:     getter desiredSize
CONSOLE MESSAGE: line 153:     getter ready
CONSOLE MESSAGE: line 153:     method abort
CONSOLE MESSAGE: line 153:     method close
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method releaseLock
CONSOLE MESSAGE: line 153:     method write
CONSOLE MESSAGE: line 153: global object
CONSOLE MESSAGE: line 153:     attribute console
CONSOLE MESSAGE: line 153:     attribute globalThis
CONSOLE MESSAGE: line 153:     method gc
CONSOLE MESSAGE: line 153:     method registerAnimator
